requirements.txt
flask
flask-cors
google-generativeai
yfinance
pandas_ta
gunicorn
matplotlib
mplfinance